+++ src/orderbook/orders/index.ts
@@ -37,6 +37,7 @@ export type OrderKind =
   | "element-erc1155"
   | "quixotic"
   | "nouns"
+  | "zora-v3"
   | "cryptopunks";
 
 // In case we don't have the source of an order readily available, we use
@@ -66,6 +67,8 @@ export const getOrderSourceByOrderKind = async (
         return sources.getOrInsert("element.market");
       case "quixotic":
         return sources.getOrInsert("quixotic.io");
+      case "zora-v3":
+        return sources.getOrInsert("zora.co");
       case "nouns":
         return sources.getOrInsert("nouns.wtf");
       default: